+++ tests/test_06_ids.py
@@ -3,23 +3,31 @@
 
 import datetime
 import filecmp
+import logging
 import time
 import zipfile
 import pytest
 import icat
+import icat.client
 import icat.config
-from icat.ids import DataSelection
+from icat.ids import IDSClient, DataSelection
 from icat.query import Query
 from conftest import DummyDatafile, UtcTimezone
 from conftest import getConfig, tmpSessionId, tmpClient
 
+logger = logging.getLogger(__name__)
 
-@pytest.fixture(scope="module")
-def client(setupicat):
-    client, conf = getConfig(ids="mandatory")
-    client.login(conf.auth, conf.credentials)
-    yield client
-    query = "SELECT df FROM Datafile df WHERE df.location IS NOT NULL"
+GiB = 1073741824
+
+class LoggingIDSClient(IDSClient):
+    """Modified version of IDSClient that logs some calls.
+    """
+    def getStatus(self, selection):
+        status = super().getStatus(selection)
+        logger.debug("getStatus(%s): %s", selection, status, stacklevel=2)
+        return status
+
+def _delete_datafiles(client, query):
     while True:
         try:
             client.deleteData(client.search(query))
@@ -27,6 +35,45 @@ def client(setupicat):
             time.sleep(10)
         else:
             break
+
+@pytest.fixture(scope="module")
+def cleanup(setupicat):
+    client, conf = getConfig(confSection="root", ids="mandatory")
+    client.login(conf.auth, conf.credentials)
+    yield
+    query = "SELECT df FROM Datafile df WHERE df.location IS NOT NULL"
+    _delete_datafiles(client, query)
+
+@pytest.fixture(scope="function")
+def client(monkeypatch, cleanup):
+    monkeypatch.setattr(icat.client, "IDSClient", LoggingIDSClient)
+    client, conf = getConfig(ids="mandatory")
+    client.login(conf.auth, conf.credentials)
+    yield client
+
+@pytest.fixture(scope="function")
+def dataset(client, cleanup_objs):
+    """A dataset to be used in the test.
+
+    The dataset will be eventually be deleted after the test.
+    """
+    inv = client.assertedSearch(Query(client, "Investigation", conditions={
+        "name": "= '10100601-ST'",
+    }))[0]
+    dstype = client.assertedSearch(Query(client, "DatasetType", conditions={
+        "name": "= 'raw'",
+    }))[0]
+    dataset = client.new("Dataset",
+                         name="e208343", complete=False,
+                         investigation=inv, type=dstype)
+    dataset.create()
+    cleanup_objs.append(dataset)
+    yield dataset
+    query = Query(client, "Datafile", conditions={
+        "dataset.id": "= %d" % dataset.id,
+        "location": "IS NOT NULL",
+    })
+    _delete_datafiles(client, query)
 
 
 # ============================ testdata ============================
@@ -397,6 +444,55 @@ def test_restore(client, case):
     # archive() applies: there is no guarantee whatsoever on the
     # outcome of the restore() call.
     print("Status of dataset %s is now %s" % (case['dsname'], status))
+
+@pytest.mark.parametrize(("case"), markeddatasets)
+def test_restoreDataCall(client, case):
+    """Test the high level call restoreData().
+
+    This is essentially a no-op as the dataset in question will
+    already be ONLINE.  It only tests that the call does not throw an
+    error.
+    """
+    dataset = getDataset(client, case)
+    client.restoreData([dataset])
+    status = client.ids.getStatus(DataSelection([dataset]))
+    assert status == "ONLINE"
+
+@pytest.mark.parametrize(("case"), markeddatasets)
+def test_restoreDataCallSelection(client, case):
+    """Test the high level call restoreData().
+
+    Same as last test, but now pass a DataSelection as argument.
+    """
+    selection = DataSelection([getDataset(client, case)])
+    client.restoreData(selection)
+    status = client.ids.getStatus(selection)
+    assert status == "ONLINE"
+
+@pytest.mark.slow
+def test_restoreData(tmpdirsec, client, dataset):
+    """Test restoring data with the high level call restoreData().
+
+    This test archives a dataset and calls restoreData() to restore it
+    again.  The size of the dataset is large enough so that restoring
+    takes some time, so that we actually can observe the call to wait
+    until the restoring is finished.  As a result, the test is rather
+    slow.  It is marked as such and thus disabled by default.
+    """
+    if not client.ids.isTwoLevel():
+        pytest.skip("This IDS does not use two levels of data storage")
+    f = DummyDatafile(tmpdirsec, "e208343.nxs", GiB)
+    query = Query(client, "DatafileFormat", conditions={
+        "name": "= 'NeXus'",
+    })
+    datafileformat = client.assertedSearch(query)[0]
+    datafile = client.new("Datafile", name=f.fname.name,
+                          dataset=dataset, datafileFormat=datafileformat)
+    client.putData(f.fname, datafile)
+    client.ids.archive(DataSelection([dataset]))
+    client.restoreData([dataset])
+    status = client.ids.getStatus(DataSelection([dataset]))
+    assert status == "ONLINE"
 
 @pytest.mark.parametrize(("case"), markeddatasets)
 def test_reset(client, case):